+++ __tests__/index.test.ts
@@ -3,3 +3,4 @@ import './wiki-links.test';
 import './alias-similarity.test';
 import './edge-cases.test';
 import './title-source.test';
+import './subheading-links.test';